# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MobileWallet backend",
    "description": "\nBackend **MobileWallet** — un rassemblement d'**agrégateurs** de paiement Mobile Money.\n\nChaque agrégateur (DigiKUNTZ aujourd'hui, d'autres demain) est un module exposant\ndeux capacités :\n\n- **navigateur IA** (`mode=browser`) : un agent Playwright + LLM pilote le checkout,\n  exécute le paiement et **déduit** le « curl replay » (requête `/charge` + `/verify`),\n  persisté comme *template* réutilisable.\n- **curl replay** (`mode=replay`) : rejoue le paiement sans navigateur via le template stocké.\n- **auto** (`mode=auto`, défaut) : tente le replay d'abord, puis bascule sur le navigateur\n  si le replay est non concluant (`fallback_browser`).\n\nToute tentative est auditée (table `transactions`) quand Supabase est configuré.\n",
    "version": "0.2"
  },
  "paths": {
...
    "/payout": {
      "post": {
        "tags": [
          "payout"
        ],
        "summary": "Effectuer un virement (payout)",
        "description": "Effectue un virement sortant débité du solde de VOTRE app.\n\nLe solde de l'app = somme de ses encaissements réussis − ses retraits. Un\nvirement supérieur au solde disponible est refusé (422) avant tout appel amont.\nLe verdict final (`successful`/`failed`/`cancelled`) est poussé via webhook +\nSocket.IO ; consultable aussi par `GET /status/{transaction_id}`.",
        "operationId": "payout_payout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agrégateur inconnu."
          },
          "400": {
            "description": "Agrégateur sans support payout."
          },
          "409": {
            "description": "Un virement est déjà en cours vers ce bénéficiaire."
          },
          "422": {
            "description": "Solde de l'app insuffisant pour ce virement."
          },
          "502": {
            "description": "Le virement n'a pas pu être initié."
          },
          "503": {
            "description": "Service de paiement amont temporairement indisponible."
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Historique des transactions",
        "description": "Dernières transactions de VOTRE app (isolation par clé API).",
        "operationId": "list_transactions_transactions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/balance": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Solde de votre app",
        "description": "Solde logique de VOTRE app = somme de vos encaissements réussis − vos retraits.\n\nC'est le montant disponible pour un virement (`POST /payout`).",
        "operationId": "app_balance_balance_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/status/{transaction_ref}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Statut d'une transaction",
        "description": "Récupère une transaction de votre app par sa référence.",
        "operationId": "transaction_status_status__transaction_ref__get",
        "parameters": [
          {
            "name": "transaction_ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Ref"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Aucune transaction pour cette référence."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_ref}/trace": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Trace IA d'une transaction (mode browser)",
        "description": "Trace tour-par-tour de l'agent IA pour une transaction browser de votre app.\n\nChaque entrée : ce que l'IA a vu (url, nb d'éléments), sa pensée, les actions\njouées, et si l'objectif a été atteint. Vide pour un paiement en mode replay.",
        "operationId": "transaction_trace_transactions__transaction_ref__trace_get",
        "parameters": [
          {
            "name": "transaction_ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Ref"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Aucune transaction pour cette référence."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_ref}/errors": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Erreurs détaillées d'une transaction (browser/replay)",
        "description": "Erreurs détaillées d'une transaction de votre app.\n\nChaque entrée précise le moteur (`engine`: browser/replay) et la SOURCE\n(`source`: ai / browser / transaction / replay) pour savoir exactement où ça\na cassé, plus la catégorie, le message clair et les détails. Vide si la\ntransaction a réussi.",
        "operationId": "transaction_errors_transactions__transaction_ref__errors_get",
        "parameters": [
          {
            "name": "transaction_ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Ref"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Aucune transaction pour cette référence."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{tx_id}/cancel": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Débloquer une transaction pending",
        "description": "Force une transaction restée **pending** à **cancelled** (de votre app).\n\nSert à débloquer une transaction coincée (ex: serveur interrompu avant le\nverdict) qui empêcherait un nouveau paiement sur le même numéro\n(`409 pending_exists`). N'agit que sur une ligne encore `pending` de votre\napp — ne réécrit jamais un verdict déjà acté.",
        "operationId": "cancel_transaction_transactions__tx_id__cancel_post",
        "parameters": [
          {
            "name": "tx_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tx Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Aucune transaction 'pending' (de votre app) avec cet id."
          },
          "503": {
            "description": "Supabase non configuré."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PayoutRequest": {
        "properties": {
          "amount": {
            "type": "integer",
            "title": "Amount",
            "description": "Montant à virer en XAF.",
            "examples": [
              5000
            ]
          },
          "account_bank_code": {
            "type": "string",
            "title": "Account Bank Code",
            "description": "Réseau du bénéficiaire — STRICT, exactement 'ORANGEMONEY' ou 'MTN' (aucune variante/conversion ; 422 sinon).",
            "examples": [
              "MTN"
            ]
          },
          "account_number": {
            "type": "string",
            "title": "Account Number",
            "description": "Numéro du compte bénéficiaire.",
            "examples": [
              "237691224472"
            ]
          },
          "receiver_name": {
            "type": "string",
            "title": "Receiver Name",
            "description": "Nom du bénéficiaire.",
            "examples": [
              "John Doe"
            ]
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Devise du virement.",
            "default": "XAF",
            "examples": [
              "XAF"
            ]
          },
          "narration": {
            "type": "string",
            "title": "Narration",
            "description": "Motif / libellé du virement.",
            "default": "",
            "examples": [
              "Paiement fournisseur"
            ]
          },
          "aggregator": {
            "type": "string",
            "title": "Aggregator",
            "description": "Nom de l'agrégateur (cf. GET /aggregators).",
            "default": "digikuntz"
          },
          "callback_url": {
            "type": "string",
            "title": "Callback Url",
            "description": "URL de callback (défaut: celle de l'app).",
            "default": ""
          },
          "end_user_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End User Ref",
            "description": "Identifiant opaque de votre utilisateur final (ex. votre user_id). Stocké tel quel pour vos rapprochements ; facultatif."
          }
        },
        "type": "object",
        "required": [
          "amount",
          "account_bank_code",
          "account_number",
          "receiver_name"
        ],
        "title": "PayoutRequest",
        "examples": [
          {
            "account_bank_code": "MTN",
            "account_number": "237691224472",
            "aggregator": "digikuntz",
            "amount": 5000,
            "currency": "XAF",
            "narration": "Paiement fournisseur",
            "receiver_name": "John Doe"
          }
        ]
      },
      "PayoutResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Vrai si le virement a abouti."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "successful | pending | failed | cancelled.",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Message clair pour l'utilisateur final.",
            "default": ""
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id",
            "description": "Référence transaction de l'agrégateur.",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Code HTTP de la réponse (200 si OK).",
            "default": 200
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "PayoutResponse",
        "description": "Vue CLIENT (dev qui intègre l'API) — l'essentiel métier.\n\nLe détail technique est réservé à l'admin (PayoutResponseDebug + ?debug=true)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé API d'app (paiements) : Authorization: Bearer <sk_…>."
      },
      "DevJWT": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT du compte developer (gestion des projets) : Authorization: Bearer <jwt>."
      },
      "AdminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key",
        "description": "Clé admin pour la supervision des developers/apps/clés."
      }
    }
  },
...
}